refactor: exported liveness functionality from index
