+++ src/common/interfaces.ts
@@ -2,3 +2,7 @@ export interface IConfig {
   get: <T>(setting: string) => T;
   has: (setting: string) => boolean;
 }
+
+export interface IServerConfig {
+  port: string;
+}
